+++ vllm_rbln/model_executor/models/optimum/sliding_window.py
@@ -19,10 +19,7 @@ from vllm.logger import init_logger
 
 from .base import ModelInputForRBLN, version_error
 from .model_base import RBLNOptimumDecoderMixin, RBLNOptimumModelBase
-from .optimum_attention_manager import AttentionManager
-from .optimum_attention_strategy import (InnerAttentionEntry,
-                                         InnerAttentionStrategy, InnerR1,
-                                         InnerR2)
+from .optimum_attention import AttentionManager, InnerAttentionEntry, InnerAttentionStrategy, InnerR1, InnerR2
 
 logger = init_logger(__name__)
